+++ src/com/intellij/idea/plugin/hybris/spring/context/CngSpringELContextsExtension.kt
@@ -53,7 +53,7 @@ class CngSpringELContextsExtension : SpringElContextsExtension() {
 
         return when {
             context is XmlText
-                && tag.localName == Labels.LABEL
+                && (tag.localName == Labels.LABEL || tag.localName == Labels.DESCRIPTION)
                 && tag.namespace == NAMESPACE_COCKPIT_NG_CONFIG_HYBRIS -> process(CngPsiHelper.resolveContextType(context), project)
 
             context is XmlAttributeValue